==> 03.notebooks/05.gemini_response_results.ipynb <==
import pyarrow as pa
import pyarrow.parquet as pq
import pandas as pd
pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None)

RESPONSE_DIR = '../00.data/01.model_response/00.gemini_response/'
RESPONSE_NAME = ['00.original_response_500.parquet',
                 '01.subject_shuffled_response_500.parquet',
                 '02.object_shuffled_response_500.parquet',
                 '03.property_scoped_subject_shuffled_response_500.parquet',
                 '04.property_scoped_object_shuffled_response_500.parquet']

LANGUAGE_LIST = ['en', 'fr', 'de', 'es', 'it', 'pt', 'ko', 'ja']

pd.set_option('display.max_rows', 10)
pd.set_option('display.max_columns', None)

tmp0 = pq.read_table(f"{RESPONSE_DIR}{RESPONSE_NAME[2]}").to_pandas()
tmp0

tmp0[['response_ko', 'response_en']]

tmp1 = pq.read_table(f"{RESPONSE_DIR}{RESPONSE_NAME[1]}").to_pandas()
tmp1[['response_ko', 'response_en']]

tmp2 = pq.read_table(f"{RESPONSE_DIR}{RESPONSE_NAME[2]}").to_pandas()
tmp2[['response_ko', 'response_en']]

tmp3 = pq.read_table(f"{RESPONSE_DIR}{RESPONSE_NAME[3]}").to_pandas()
tmp3[['response_ko', 'response_en']]
