==> 03.notebooks/06.HFmodels_download.ipynb <==
from transformers import AutoModelForCausalLM, AutoTokenizer

MODEL_DIR = '../01.models/'

def download_model(model_id: str, save_dir: str):
    print(f"📥 Downloading: {model_id} → {save_dir}")
    tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=True)
    model = AutoModelForCausalLM.from_pretrained(model_id, trust_remote_code=True)
    tokenizer.save_pretrained(save_dir)
    model.save_pretrained(save_dir)
    print(f"Saved: {model_id} in {save_dir}")


download_model('meta-llama/Llama-3.1-8B-Instruct', MODEL_DIR + 'Meta/meta-llama/Llama-3.1-8B-Instruct')
download_model('mistralai/Mistral-Nemo-Instruct-2407', MODEL_DIR + 'Mistral/mistralai/Mistral-Nemo-Instruct-2407')
download_model('Qwen/Qwen3-8B', MODEL_DIR + 'Qwen/Qwen3-8B')
download_model('Qwen/Qwen3-14B', MODEL_DIR + 'Qwen/Qwen3-14B')

download_model('meta-llama/Llama-3.1-8B-Instruct', MODEL_DIR + 'Meta/meta-llama/Llama-3.1-8B-Instruct')
