==> 03.notebooks/08.HFmodels_response_result.ipynb <==
import pyarrow as pa
import pyarrow.parquet as pq

TEMPLATE_DIR = '../00.data/00.wikidata/03.wikidata_template/'
TEMPLATE_NAME = ['00.original_template_100.parquet',
                 '01.subject_shuffled_template_100.parquet',
                 '02.object_shuffled_template_100.parquet',
                 '03.property_scoped_subject_shuffled_template_100.parquet',
                 '04.property_scoped_object_shuffled_template_100.parquet']

MODEL_DIR = '../01.models/'
MODEL_NAME = ['Meta/meta-llama/Llama-3.1-8B-Instruct/',
              'Mistral/mistralai/Mistral-Nemo-Instruct-2407/',
              'Qwen/Qwen3-8B/',
              'Qwen/Qwen3-14B/']
RESPONSE_DIR = '../00.data/01.model_response/01.HF_Models_response/'

LANGUAGE_LIST = ['en', 'fr', 'de', 'es', 'it', 'pt', 'ko', 'ja']

RESPONSE_NAME = ['00.original_response_100.parquet',
                 '01.subject_shuffled_response_100.parquet',
                 '02.object_shuffled_response_100.parquet',
                 '03.property_scoped_subject_shuffled_response_100.parquet',
                 '04.property_scoped_object_shuffled_response_100.parquet']


import pandas as pd
pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None)

tmp0 = pq.read_table(f"../00.data/01.model_response/01.HF_Models_response/Qwen/Qwen3-14B/00.original_response_100.parquet").to_pandas()
tmp0

tmp0 = pq.read_table(f"{RESPONSE_DIR}{MODEL_NAME[2]}{RESPONSE_NAME[0]}").to_pandas()

tmp0
